# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Job Website API",
        "description": "Job Website API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/job-locations/{id}": {
            "get": {
                "tags": [
                    "Job Locations"
                ],
                "summary": "Get job location by id",
                "operationId": "94a7913aec898b82010cc6b3d1496b6e",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "description": "application/json",
                        "required": false
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Job Location ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved job location",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "error": false,
                                    "message": "Xử lí thành công",
                                    "data": {
                                        "job_location": {
                                            "current_page": 1,
                                            "data": [
                                                {
                                                    "id": 3,
                                                    "job_id": 3,
                                                    "location": "Hà Nội: 84 Tô Ngọc Vân"
                                                }
                                            ],
                                            "first_page_url": "http://localhost:8000/api/job-locations/3?page=1",
                                            "from": 1,
                                            "last_page": 1,
                                            "last_page_url": "http://localhost:8000/api/job-locations/3?page=1",
                                            "links": [
                                                {
                                                    "url": null,
                                                    "label": "&laquo; Previous",
                                                    "active": false
                                                },
                                                {
                                                    "url": "http://localhost:8000/api/job-locations/3?page=1",
                                                    "label": "1",
                                                    "active": true
                                                },
                                                {
                                                    "url": null,
                                                    "label": "Next &raquo;",
                                                    "active": false
                                                }
                                            ],
                                            "next_page_url": null,
                                            "path": "http://localhost:8000/api/job-locations/3",
                                            "per_page": 1,
                                            "prev_page_url": null,
                                            "to": 1,
                                            "total": 1
                                        }
                                    },
                                    "status_code": 200
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                }
            },
            "put": {
                "tags": [
                    "Job Locations"
                ],
                "summary": "Update a job location",
                "operationId": "29d03242bbf54b2f0cd2b35079408b0e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Job location id",
                        "required": true
                    },
                    {
                        "name": "Accept",
                        "in": "header",
                        "description": "application/json",
                        "required": false
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer {token}",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "job_id": 1,
                                "location": "abc"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successfully updated job location",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "error": false,
                                    "message": "Xử lí thành công",
                                    "data": {
                                        "job_location": {
                                            "id": 57,
                                            "job_id": "1",
                                            "location": "abcdef"
                                        }
                                    },
                                    "status_code": 200
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Job Locations"
                ],
                "summary": "Delete a job location",
                "operationId": "1d5af4ad2cef2a671d69a8a33c2b2998",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Job location id",
                        "required": true
                    },
                    {
                        "name": "Accept",
                        "in": "header",
                        "description": "application/json",
                        "required": false
                    },
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer {token}",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully deleted job location",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "error": false,
                                    "message": "Xử lí thành công",
                                    "data": {
                                        "job_location": {
                                            "id": 57,
                                            "job_id": 1,
                                            "location": "abcdef"
                                        }
                                    },
                                    "status_code": 200
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                }
            }
        },
        "/api/job-locations/job/{job_id}": {
            "get": {
                "tags": [
                    "Job Locations"
                ],
                "summary": "Get job locations by job id",
                "operationId": "699f5fb8e3e71c7360ddbd25073c6d6c",
                "parameters": [
                    {
                        "name": "Accept",
                        "in": "header",
                        "description": "application/json",
                        "required": false
                    },
                    {
                        "name": "count_per_page",
                        "in": "query",
                        "description": "Number of job locations per page",
                        "required": false
                    },
                    {
                        "name": "job_id",
                        "in": "path",
                        "description": "Job ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved job locations",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "error": false,
                                    "message": "Xử lí thành công",
                                    "data": {
                                        "job_location": {
                                            "current_page": 1,
                                            "data": [
                                                {
                                                    "id": 1,
                                                    "job_id": 1,
                                                    "location": "Hồ Chí Minh: 72 Bình Giã"
                                                }
                                            ],
                                            "first_page_url": "http://localhost:8000/api/job-locations/job/1?page=1",
                                            "from": 1,
                                            "last_page": 1,
                                            "last_page_url": "http://localhost:8000/api/job-locations/job/1?page=1",
                                            "links": [
                                                {
                                                    "url": null,
                                                    "label": "&laquo; Previous",
                                                    "active": false
                                                },
                                                {
                                                    "url": "http://localhost:8000/api/job-locations/job/1?page=1",
                                                    "label": "1",
                                                    "active": true
                                                },
                                                {
                                                    "url": null,
                                                    "label": "Next &raquo;",
                                                    "active": false
                                                }
                                            ],
                                            "next_page_url": null,
                                            "path": "http://localhost:8000/api/job-locations/job/1",
                                            "per_page": 15,
                                            "prev_page_url": null,
                                            "to": 1,
                                            "total": 1
                                        }
                                    },
                                    "status_code": 200
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                }
            }
        },
...
    },
    "components": {
        "responses": {
...
            "NotFound": {
                "description": "Không tìm thấy",
                "content": {
                    "application/json": {
                        "schema": {},
                        "example": {
                            "error": true,
                            "message": "Không tìm thấy",
                            "data": null,
                            "status_code": 404
                        }
                    }
                }
            },
...
        }
    }
}